support bulk termination of workflows (#2176)
* support bulk termination of workflows

* move BulkResponse to common
